Consolidate loggger to agent
